Add files via upload
You are given all numbers between 1 and n but there is one number missing.
Your task is to write a program to find and print the missing number. The
program will accept input on two lines, the first line will accept the value of n
and the second line will accept n−1 numbers, each between 1 and n. The output
of the program will be the missing number. Note that the program should run
in under 1 s.
Input:
10
1 2 3 4 5 6 7 8 10
Output: